+++ src/ui/activities/auth/mod.rs
@@ -29,7 +29,8 @@ const RADIO_PROTOCOL_SCP: usize = 1;
 const RADIO_PROTOCOL_FTP: usize = 2;
 const RADIO_PROTOCOL_FTPS: usize = 3;
 const RADIO_PROTOCOL_S3: usize = 4;
-const RADIO_PROTOCOL_SMB: usize = 5;
+const RADIO_PROTOCOL_WEBDAV: usize = 5;
+const RADIO_PROTOCOL_SMB: usize = 6;
 
 // -- components
 #[derive(Debug, Eq, PartialEq, Clone, Hash)]
@@ -71,6 +72,7 @@ pub enum Id {
     Title,
     Username,
     WaitPopup,
+    WebDAVUri,
     WindowSizeError,
 }
 
@@ -155,6 +157,8 @@ pub enum UiMsg {
     ShowSaveBookmarkPopup,
     UsernameBlurDown,
     UsernameBlurUp,
+    WebDAVUriBlurDown,
+    WebDAVUriBlurUp,
     WindowResized,
 }
 
@@ -164,6 +168,7 @@ enum InputMask {
     Generic,
     AwsS3,
     Smb,
+    WebDAV,
 }
 
 // Store keys
@@ -240,6 +245,7 @@ impl AuthActivity {
             | FileTransferProtocol::Scp
             | FileTransferProtocol::Sftp => InputMask::Generic,
             FileTransferProtocol::Smb => InputMask::Smb,
+            FileTransferProtocol::WebDAV => InputMask::WebDAV,
         }
     }
 }