+++ CHANGELOG.md
@@ -43,8 +43,12 @@ Released on
   - Changed `-t` to `theme`
   - Changed `-u` to `update`
   - Changed `-c` to `config`
+- Introduced support for [WebDAV](https://www.rfc-editor.org/rfc/rfc4918)
+  - It is now possible also to connect directly to WebDAV server with the syntax `http(s)://username:password@google.com`
 - Bugfix:
   - [Issue 232](https://github.com/veeso/termscp/issues/232): AWS S3 wasn't working anymore due to rust-s3 outdate
+- Dependencies:
+  - Added `remotefs-webdav 0.1.1`
 
 ## 0.12.3